refactor(act): move public builders into src/builders/
Group the four fluent builders (state, slice, projection, act) under
a dedicated builders/ subfolder with its own barrel. Clarifies the
codebase layout — internal pipelines live in internal/, public
builders in builders/, orchestrator at the root.

Pure file move + import-path updates. No behavioral or API changes.
Public surface unchanged: src/index.ts re-exports from
./builders/index.ts so consumers keep importing from
"@rotorsoft/act".
