+++ libs/act/src/index.ts
@@ -5,13 +5,10 @@ import "./signals.js";
  * @module act
  * Main entry point for the Act framework. Re-exports all core APIs.
  */
-export * from "./act-builder.js";
 export * from "./act.js";
 export * from "./adapters/index.js";
+export * from "./builders/index.js";
 export * from "./config.js";
 export * from "./ports.js";
-export * from "./projection-builder.js";
-export * from "./slice-builder.js";
-export * from "./state-builder.js";
 export * from "./types/index.js";
 export * from "./utils.js";
