+++ libs/act/test/tracing.spec.ts
@@ -1,11 +1,11 @@
 import { z } from "zod";
 import { InMemoryCache } from "../src/adapters/InMemoryCache.js";
 import { InMemoryStore } from "../src/adapters/InMemoryStore.js";
+import { state } from "../src/builders/state-builder.js";
 import * as drain from "../src/internal/drain.js";
 import * as es from "../src/internal/event-sourcing.js";
 import { buildDrain, buildEs } from "../src/internal/tracing.js";
 import { cache, log, store } from "../src/ports.js";
-import { state } from "../src/state-builder.js";
 import type { Logger } from "../src/types/index.js";
 import { ZodEmpty } from "../src/types/schemas.js";
